move http_call_indirection.rs and network_state.rs to types crate
