+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod constants;
 pub mod fetch;
-pub mod http_call_indirection;
 pub mod init_tunnel;
-pub mod network_state;
 mod utils;
+pub mod types;
