+++ src/types/network_state.rs
